refactoring of cli folders
