--- supervisely/cli/download/__init__.py
@@ -1,1 +1,0 @@
-from supervisely.cli.download.download import download_run
