+++ supervisely/cli/project/__init__.py
@@ -1,0 +1,3 @@
+from supervisely.cli.project.project_get import get_project_name_run
+
+from supervisely.cli.project.project_download import download_run
