--- supervisely/cli/remove/__init__.py
@@ -1,1 +1,0 @@
-from supervisely.cli.remove.remove import remove_file_run, remove_dir_run
