+++ supervisely/cli/teamfiles/__init__.py
@@ -1,0 +1,3 @@
+from supervisely.cli.teamfiles.teamfiles_remove import remove_file_run, remove_dir_run
+
+from supervisely.cli.teamfiles.teamfiles_upload import upload_to_teamfiles_run
