+++ supervisely/cli/teamfiles/teamfiles_remove.py
