--- supervisely/cli/upload/__init__.py
@@ -1,1 +1,0 @@
-from supervisely.cli.upload.upload import upload_to_teamfiles_run, set_task_output_dir_run
